# Conflicts:
#	src/main/java/com/siyan1234/itproject2nd/map/client/KakaoMapClient.java
#	src/main/java/com/siyan1234/itproject2nd/map/controller/KakaoMapApiController.java
#	src/main/java/com/siyan1234/itproject2nd/map/controller/KakaoMapPageController.java
#	src/main/java/com/siyan1234/itproject2nd/map/service/KakaoMapService.java
#	src/main/java/com/siyan1234/itproject2nd/map/support/KakaoMapApiProperties.java
#	src/main/java/com/siyan1234/itproject2nd/map/support/KakaoMapApiUrls.java
#	src/main/java/com/siyan1234/itproject2nd/map/support/KakaoMapMessages.java
#	src/main/resources/application.yaml
#	src/main/resources/static/css/kakao-map.css
#	src/main/resources/static/js/kakao-map.js
#	src/main/resources/templates/map/kakao-map.html

+++ src/main/java/com/siyan1234/itproject2nd/stock/dto/StockDto.java
@@ -15,8 +15,8 @@ public class StockDto {
     private int id;
     private String symbol;
     private String name;
-    private long price;
-    private long changePrice;
+    private double price;
+    private double changePrice;
     private double changeRate;
     private List<StockPointDto> points;
 }
